Reworked array of pppd args (#295)
This is a re-worked variant for commit de87e36c4bf4793274c650df4cb8bef726876434
which had to be reverted for the patch series to apply.

Original description:

Use a dynamically (re)allocated array instead of a fixed-length args[].

This eliminates the need to carefully match the number of NULL
placeholders with the number of possible options, and also the need for
a-posteriori check of bounds.